+++ src/mesido/financial_mixin.py
@@ -975,16 +975,14 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
             else:
                 price_profile = Timeseries(self.times(), np.zeros(len(self.times())))
 
-            sum = 0.0
-            for i in range(1, len(self.times())):
-                sum += (
-                    variable_operational_cost_coefficient
-                    * (heat_charge[i] + heat_discharge[i])
-                    * timesteps[i - 1]
-                )
-                sum += price_profile.values[i] * pump_power[i] * timesteps[i - 1] / eff
+            sum_ = ca.sum1(
+                variable_operational_cost_coefficient
+                * (heat_charge[1:] + heat_discharge[1:])
+                * timesteps
+            )
+            sum_ += ca.sum1(price_profile.values[1:] * pump_power[1:] * timesteps / eff)
 
-            constraints.append(((variable_operational_cost - sum) / nominal, 0.0, 0.0))
+            constraints.append(((variable_operational_cost - sum_) / nominal, 0.0, 0.0))
 
         for asset in [
             *self.energy_system_components.get("pump", []),
@@ -1011,11 +1009,9 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
             else:
                 price_profile = Timeseries(self.times(), np.zeros(len(self.times())))
 
-            sum = 0.0
-            for i in range(1, len(self.times())):
-                sum += price_profile.values[i] * pump_power[i] * timesteps_hr[i - 1] / eff
+            sum_ = ca.sum1(price_profile.values[1:] * pump_power[1:] * timesteps_hr / eff)
 
-            constraints.append(((variable_operational_cost - sum) / nominal, 0.0, 0.0))
+            constraints.append(((variable_operational_cost - sum_) / nominal, 0.0, 0.0))
 
         for s in self.energy_system_components.get("heat_source", []):
             heat_source = self.__state_vector_scaled(f"{s}.Heat_source", ensemble_member)
@@ -1072,17 +1068,14 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
                 )  # [W]
             else:
                 nominator_vector = heat_source
-            sum = 0.0
-            for i in range(1, len(self.times())):
-                sum += (
-                    variable_operational_cost_coefficient
-                    * nominator_vector[i]
-                    * timesteps_hr[i - 1]
-                    / denominator
-                )
-                sum += price_profile.values[i] * pump_power[i] * timesteps_hr[i - 1] / eff
 
-            constraints.append(((variable_operational_cost - sum) / nominal, 0.0, 0.0))
+            sum_ = (
+                ca.sum1(variable_operational_cost_coefficient * nominator_vector[1:] * timesteps_hr)
+                / denominator
+            )
+            sum_ += ca.sum1(price_profile.values[1:] * pump_power[1:] * timesteps_hr / eff)
+
+            constraints.append(((variable_operational_cost - sum_) / nominal, 0.0, 0.0))
 
         for hp in [
             *self.energy_system_components.get("heat_pump", []),
@@ -1111,20 +1104,15 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
             else:
                 price_profile = Timeseries(self.times(), np.zeros(len(self.times())))
 
-            sum = 0.0
-            for i in range(1, len(self.times())):
-                sum += (
-                    variable_operational_cost_coefficient
-                    * elec_consumption[i]
-                    * timesteps_hr[i - 1]
-                )
-                sum += price_profile.values[i] * pump_power[i] * timesteps_hr[i - 1] / eff
-                if hp not in self.energy_system_components.get("heat_pump_elec", []):
-                    # assuming that if heatpump has electricity port, the cost for the electricity
-                    # are already made by the electricity producer and transport
-                    sum += price_profile.values[i] * elec_consumption[i] * timesteps_hr[i - 1]
-
-            constraints.append(((variable_operational_cost - sum) / nominal, 0.0, 0.0))
+            sum_ = ca.sum1(
+                variable_operational_cost_coefficient * elec_consumption[1:] * timesteps_hr
+            )
+            sum_ += ca.sum1(price_profile.values[1:] * pump_power[1:] * timesteps_hr / eff)
+            if hp not in self.energy_system_components.get("heat_pump_elec", []):
+                # assuming that if heatpump has electricity port, the cost for the electricity
+                # are already made by the electricity producer and transport
+                sum_ += ca.sum1(price_profile.values[1:] * elec_consumption[1:] * timesteps_hr)
+            constraints.append(((variable_operational_cost - sum_) / nominal, 0.0, 0.0))
 
         for ac in self.energy_system_components.get("airco", []):
             heat_airco = self.__state_vector_scaled(f"{ac}.Heat_airco", ensemble_member)
@@ -1136,11 +1124,9 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
             variable_operational_cost_coefficient = parameters[
                 f"{ac}.variable_operational_cost_coefficient"
             ]
-            sum = 0.0
-            for i in range(1, len(self.times())):
-                sum += variable_operational_cost_coefficient * heat_airco[i] * timesteps_hr[i - 1]
+            sum_ = ca.sum1(variable_operational_cost_coefficient * heat_airco[1:] * timesteps_hr)
 
-            constraints.append(((variable_operational_cost - sum) / nominal, 0.0, 0.0))
+            constraints.append(((variable_operational_cost - sum_) / nominal, 0.0, 0.0))
 
         for demand in self.energy_system_components.get("gas_demand", []):
             gas_mass_flow = self.__state_vector_scaled(
@@ -1156,13 +1142,8 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
                 f"{demand}.variable_operational_cost_coefficient"
             ]
 
-            sum = 0.0
-            for i in range(1, len(self.times())):
-                sum += (
-                    variable_operational_cost_coefficient * gas_mass_flow[i] * timesteps_hr[i - 1]
-                )
-
-            constraints.append(((variable_operational_cost - sum) / nominal, 0.0, 0.0))
+            sum_ = ca.sum1(variable_operational_cost_coefficient * gas_mass_flow[1:] * timesteps_hr)
+            constraints.append(((variable_operational_cost - sum_) / nominal, 0.0, 0.0))
 
         for gs in self.energy_system_components.get("gas_source", []):
             gas_produced_g_s = self.__state_vector_scaled(
@@ -1177,14 +1158,11 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
                 f"{gs}.variable_operational_cost_coefficient"
             ]
             timesteps_sec = np.diff(self.times())
-            sum = 0.0
-            for i in range(1, len(self.times())):
-                sum += (
-                    variable_operational_cost_coefficient
-                    * gas_produced_g_s[i]
-                    * timesteps_sec[i - 1]
-                )  # [euro/g] * [g/s] * [s]
-            constraints.append(((variable_operational_cost - sum) / nominal, 0.0, 0.0))
+            sum_ = ca.sum1(
+                variable_operational_cost_coefficient * gas_produced_g_s[1:] * timesteps_sec
+            )
+            # [euro/g] * [g/s] * [s]
+            constraints.append(((variable_operational_cost - sum_) / nominal, 0.0, 0.0))
 
         for es in self.energy_system_components.get("electricity_source", []):
             elec_produced_w = self.__state_vector_scaled(
@@ -1198,12 +1176,10 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
             variable_operational_cost_coefficient = parameters[  # euro / Wh
                 f"{es}.variable_operational_cost_coefficient"
             ]
-            sum = 0.0
-            for i in range(1, len(self.times())):
-                sum += (
-                    variable_operational_cost_coefficient * elec_produced_w[i] * timesteps_hr[i - 1]
-                )  # [euro/Wh] * [W] * [hr]
-            constraints.append(((variable_operational_cost - sum) / nominal, 0.0, 0.0))
+            sum_ = ca.sum1(
+                variable_operational_cost_coefficient * elec_produced_w[1:] * timesteps_hr
+            )  # [euro/Wh] * [W] * [hr]
+            constraints.append(((variable_operational_cost - sum_) / nominal, 0.0, 0.0))
 
         # for a in self.heat_network_components.get("ates", []):
         # TODO: needs to be replaced with the positive or abs value of this, see varOPEX,
@@ -1242,15 +1218,13 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
                 f"{electrolyzer}.variable_operational_cost_coefficient"
             ]
 
-            sum = 0.0
-            for i in range(1, len(self.times())):
-                sum += (
-                    variable_operational_cost_coefficient
-                    * power_consumer[i]
-                    * timesteps_hr[i - 1]  # gas_mass_flow unit is g/s
-                )
+            sum_ = ca.sum1(
+                variable_operational_cost_coefficient
+                * power_consumer[1:]
+                * timesteps_hr  # gas_mass_flow unit is g/s
+            )
 
-            constraints.append(((variable_operational_cost - sum) / nominal, 0.0, 0.0))
+            constraints.append(((variable_operational_cost - sum_) / nominal, 0.0, 0.0))
 
         # for a in self.heat_network_components.get("ates", []):
         # TODO: needs to be replaced with the positive or abs value of this, see varOPEX,
@@ -1636,7 +1610,12 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
                     carrier_name = attr["name"]
                     cost_multiplier = 1.0  # priceprofile gas is in EUR/g
             if carrier_name is not None:
-                price_profile = self.get_timeseries(f"{carrier_name}.price_profile").values
+                price_profile_timeseries = self.get_timeseries(f"{carrier_name}.price_profile")
+                # The slicing is required if the timeseries wasn't adapted in the read
+                mask = (price_profile_timeseries.times >= self.times()[0]) & (
+                    price_profile_timeseries.times <= self.times()[-1]
+                )
+                price_profile = price_profile_timeseries.values[mask]
 
                 if demand in self.energy_system_components.get("gas_demand", []):
                     energy_flow = self.__state_vector_scaled(
@@ -1652,12 +1631,10 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
                 variable_revenue = self.extra_variable(variable_revenue_var, ensemble_member)
                 nominal = self.variable_nominal(variable_revenue_var)
 
-                sum = 0.0
                 timesteps = np.diff(self.times()) * cost_multiplier
-                for i in range(1, len(self.times())):
-                    sum += price_profile[i] * energy_flow[i] * timesteps[i - 1]
+                sum_ = ca.sum1(price_profile[1:] * energy_flow[1:] * timesteps)
 
-                constraints.append(((variable_revenue - sum) / (nominal), 0.0, 0.0))
+                constraints.append(((variable_revenue - sum_) / (nominal), 0.0, 0.0))
 
         return constraints